requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
fuzzywuzzy==0.18.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Levenshtein==0.26.0
MarkupSafe==3.0.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.2
qrcode==8.0
RapidFuzz==3.10.0
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
zebra==0.1.0
zipp==3.20.2
